+++ database/migrations/2019_09_27_165117_create_reference_bank_table.php
@@ -17,7 +17,7 @@ class CreateReferenceBankTable extends Migration
             $table->increments('id');
             $table->string('ref',20);
             $table->string('bank',20);
-            $table->float('amount',8);
+            $table->decimal('amount',12,2);
             $table->timestamps();
         });
     }
